code/requirements.txt
requests>=2.28.0
meteostat>=1.6.0
pandas>=1.5.0
python-dotenv>=0.19.0